Move component, context, and lib files out of app folder
